Fix fanotify_supported()
Parse negative value of --timeout argument and error only after setting
up all watches.

this is used to improve the helper to detect support for fanotify on a
specific filesystem (e.g. overlayfs) and test permission for specific
flags (e.g. --filesystem).

With this improvement, we can remove the workaround for not testing
fanotify inside containers, because the helper will detect if kernel
supports fanotify support inside containers or not (overlayfs support
for fanotify should be supported in future kernels).

+++ src/common.cpp
@@ -197,7 +197,7 @@ void warn_inotify_init_error(int fanotify) {
 	}
 }
 
-bool is_timeout_option_valid(unsigned int* timeout, char* o) {
+bool is_timeout_option_valid(long* timeout, char* o) {
 	if ((o == NULL) || (*o == '\0')) {
 		fprintf(stderr,
 			"The provided value is not a valid timeout value.\n"
